+++ cookie-cutter-template/cookiecutter.json
@@ -1,11 +1,12 @@
 {
   "repo_name": "Project",
-  "__repo_name": "{{cookiecutter.repo_name|lower|replace(' ', '_')}}",
   "package_name": "Project",
-  "__package_name": "{{cookiecutter.package_name|lower|replace(' ', '_')}}",
   "description": "...",
   "autor_full_name": "...",
   "autor_email": "...",
-  "year": "{% now 'utc', '%Y' %}",
+  "error_tags": "",
+  "license_year": "{% now 'utc', '%Y' %}",
+  "__repo_name_slug": "{{cookiecutter.repo_name|lower|replace(' ', '_')}}",
+  "__package_name_slug": "{{cookiecutter.package_name|lower|replace(' ', '_')}}",
   "_extensions": ["cookiecutter.extensions.TimeExtension"]
 }
