+++ cookie-cutter-template/{{cookiecutter.__repo_name_slug}}/.gitignore
