+++ cookie-cutter-template/{{cookiecutter.__repo_name_slug}}/.pre-commit-config.yaml
