+++ cookie-cutter-template/{{cookiecutter.__repo_name_slug}}/LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) {{cookiecutter.year}} Exasol
+Copyright (c) {{cookiecutter.license_year}} Exasol
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
