+++ cookie-cutter-template/{{cookiecutter.__repo_name_slug}}/README.md
@@ -1,0 +1,37 @@
+<h1 align="center">{{cookiecutter.repo_name}}</h1>
+
+<p align="center">
+{{cookiecutter.description}}
+</p>
+
+<p align="center">
+
+<a href="https://opensource.org/licenses/MIT">
+    <img src="https://img.shields.io/pypi/l/{{cookiecutter.__repo_name_slug}}" alt="License">
+</a>
+<a href="https://pypi.org/project/{{cookiecutter.__repo_name_slug}}/">
+    <img src="https://img.shields.io/pypi/dm/{{cookiecutter.__repo_name_slug}}" alt="Downloads">
+</a>
+<a href="https://pypi.org/project/{{cookiecutter.__repo_name_slug}}/">
+    <img src="https://img.shields.io/pypi/pyversions/{{cookiecutter.__repo_name_slug}}" alt="Supported Python Versions">
+</a>
+<a href="https://pypi.org/project/{{cookiecutter.__repo_name_slug}}/">
+    <img src="https://img.shields.io/pypi/v/{{cookiecutter.__repo_name_slug}}" alt="PyPi Package">
+</a>
+</p>
+
+## 🚀 Features
+
+## 🔌️ Prerequisites
+
+- [Python](https://www.python.org/) >= 3.9
+
+## 💾 Installation
+
+```shell
+pip install {{cookiecutter.__repo_name_slug}}
+```
+
+## 📚 Documentation
+
+For futher details, checkout the latest [documentation](https://exasol.github.io/python-toolbox/).
