+++ cookie-cutter-template/{{cookiecutter.__repo_name_slug}}/error_code_config.yml
@@ -1,0 +1,2 @@
+error-tags:
+  {{cookiecutter.error_tags}}
