+++ cookie-cutter-template/{{cookiecutter.__repo_name_slug}}/noxconfig.py
@@ -1,0 +1,29 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+from pathlib import Path
+from tempfile import TemporaryDirectory
+from typing import (
+    Any,
+    Iterable,
+    MutableMapping,
+)
+
+from nox import Session
+
+
+@dataclass(frozen=True)
+class Config:
+    root: Path = Path(__file__).parent
+    doc: Path = Path(__file__).parent / "doc"
+    version_file: Path = Path(__file__).parent / "exasol" / "{{cookiecutter.__package_name_slug}}" / "version.py"
+    path_filters: Iterable[str] = (
+        "dist",
+        ".eggs",
+        "venv",
+    )
+
+    plugins = []
+
+
+PROJECT_CONFIG = Config()
