+++ cookie-cutter-template/{{cookiecutter.__repo_name_slug}}/test/integration/integration_smoke_test.py
@@ -1,0 +1,2 @@
+def test_unit_smoke_test():
+    assert True
