+++ cookie-cutter-template/{{cookiecutter.__repo_name_slug}}/test/unit/unit_smoke_test.py
