--- cookie-cutter-template/{{cookiecutter.__repo_name}}/README.md
@@ -1,25 +1,0 @@
-Exasol {{cookiecutter.package_name.capitalize()}}
-###############
-
-
-Exasol {{cookiecutter.package_name.capitalize()}} is a python library to ... .
-
-🚀 Features
-------------
-
-🔌️ Prerequisites
------------------
-
-- `Python <https://www.python.org/>`_ >= 3.8
-
-💾 Installation
-----------------
-
-.. code-block:: shell
-
-    pip install exasol-{{cookiecutter.package_name}}
-
-📚 Documentation
-----------------
-
-The latest documentation can be found `here <https://exasol.github.io/{{cookiecutter.repo_name}}/>`_.
