--- cookie-cutter-template/{{cookiecutter.__repo_name}}/error_code_config.yml
@@ -1,3 +1,0 @@
-error-tags:
-  BFSP:
-    highest-index: 0
