--- cookie-cutter-template/{{cookiecutter.__repo_name}}/test/conftest.py
@@ -1,11 +1,0 @@
-import os
-
-import pytest
-
-
-@pytest.fixture
-def temporary_directory(tmp_path):
-    old = os.getcwd()
-    os.chdir(tmp_path)
-    yield tmp_path
-    os.chdir(old)
